+++ packages/core/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./format.js";
 export * from "./hash.js";
 export * from "./schemas.js";
+export * from "./title.js";
 export * from "./validation.js";